Added i2c_Test project
This project includes an i2c module as well as a module that is supposed to test this (and so we learn how to use it). It will do this by connecting the Temperature dev board, the Altimeter dev board, and the RTC dev board on the same i2c bus connected to the FPGA dev board. The pins are pio1 for SDA and pio2 for SCL. As of pushing this it has not been tested, but a bitstream has been run successully.