Version V3.0 of I2C module (works)
This i2c module works and takes data from temp, alt and rtc. The difference is that IOBUFs is implemented in order to achieve a high impedence state at the IO pins to the FPGA. The i2c master is modified for this purpose and it needs the top-level wrapper to instantiate IOBUFFERS for this. (So in order for the module to work, it needs the wrapper to look like it does).

I checked the temperature data and it gave 23 degrees so at least that looks good (I have not checked if the values from alt and rtc are reasonably)