Fixed issues with reading RTC and ALT
The RTC and ALT output 3 bytes, but the previous version only read 2 bytes. This is now fixed. The only issue that remains is the the data read from ALT is all '1's. I assume there is a problem with the addressing so that the data being read is just the line being held high.